backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
pymongo==4.5.0
pydantic>=2.6.4
python-dotenv>=1.0.1
python-jose>=3.3.0
python-multipart>=0.0.9
bcrypt>=4.0.0
PyJWT>=2.10.1
passlib>=1.7.4
cryptography>=42.0.8
requests>=2.31.0
httpx>=0.25.0
aioredis>=2.0.0
redis>=5.0.0